Improve robustness, especially regarding file permission errors

+++ src/build.ts
@@ -2,6 +2,7 @@ import { SourceFile, StaticFile, getSourceFiles, getStaticFiles, resolveSourceFi
 import fs from "fs/promises";
 import path from "path";
 import colors from "colors"
+import { waitTimeout } from "./util.js";
 
 export async function build(srcDir: string, staticDir: string, distDir: string) {
 	console.log(colors.cyan('Building output directory...'))
@@ -19,6 +20,26 @@ export async function build(srcDir: string, staticDir: string, distDir: string) 
 	}
 }
 
+/**
+ * Will try to access all files given as paths to make sure that they can be used.
+ * @returns a promise holding a boolean that is true if the files are ok and false if they are not.
+ */
+async function accessFiles(paths: string[] | string, timeoutMs = 100): Promise<boolean> {
+	await waitTimeout(timeoutMs)
+	
+	if (typeof paths === "string")
+		paths = [paths]
+	
+	let pathsAreOkPromises = paths.map(
+		async path => fs.access(path).then(() => true).catch(() => false)
+	)
+	
+	let pathsAreOk = await Promise.all(pathsAreOkPromises)
+	let containsErroringFile = pathsAreOk.includes(false)
+	
+	return !containsErroringFile
+}
+
 export async function buildSourceFile(source: SourceFile, distDir: string) {
 	if (!source.isPublic)
 		return
@@ -28,13 +49,15 @@ export async function buildSourceFile(source: SourceFile, distDir: string) {
 	
 	if (source.isDeleted) {
 		// delete destination file
-		
-		try {
-			await fs.unlink(destinationPath)
-		} catch (e) {
-			console.error("Could not delete file " + destinationPath)
-			console.error(e)
-		}
+		if (await accessFiles(destinationPath) == true)
+			try {
+				fs.unlink(destinationPath)
+			} catch (e) {
+				console.error("Could not delete file " + destinationPath)
+				console.error(e)
+			}
+		else
+			console.error("Cannot delete file " + destinationPath + ", not enough permissions or file does not exist.")
 		
 		return
 	}
@@ -42,6 +65,10 @@ export async function buildSourceFile(source: SourceFile, distDir: string) {
 	// compose destination file out of dependency chain
 	let files = [...source.dependencies, source]
 	let content = ""
+	
+	if (await accessFiles(files.map(file => file.totalPath)) == false) {
+		console.error(`Cannot update file ${destinationPath}, not enough permission or source files do not exist.`)
+	}
 	
 	for (const file of files) {
 		let dependencyContent = prepareContent(await fs.readFile(file.totalPath, 'utf8'), path.join(distDir, file.relativePath), destinationPath)
@@ -95,6 +122,10 @@ function prepareContent(content: string, filePath: string, destinationPath: stri
 export async function buildStaticFile(file: StaticFile, distDir: string) {
 	let destinationPath = path.join(distDir, file.relativePath)
 	let destinationDir = path.dirname(destinationPath)
+	
+	if (await accessFiles(file.totalPath) == false) {
+		console.error(`Cannot update ${destinationPath}, not enough permission or source file does not exist.`)
+	}
 	
 	await fs.mkdir(destinationDir, { recursive: true })
